% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_extract_dbms_data}
\alias{analyse_extract_dbms_data}
\title{FUNCTION: analyse_extract_dbms_data}
\usage{
analyse_extract_dbms_data(d, dbms)
}
\description{
Perform a single Wilcoxon rank-sum test for two input data sets.
}

