% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_vargha_delaney_effect_size}
\alias{analyse_vargha_delaney_effect_size}
\title{FUNCTION: analyse_vargha_delaney_effect_size}
\usage{
analyse_vargha_delaney_effect_size(d)
}
\description{
Perform an effect size analysis of the timings for the standard and virtual mutation analysis techniques.
By default, specify that the calculation should use the effectsize_accurate function for calculating the statistic.
}

