% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_vargha_delaney_effect_size_specify}
\alias{analyse_vargha_delaney_effect_size_specify}
\title{FUNCTION: analyse_vargha_delaney_effect_size_specify}
\usage{
analyse_vargha_delaney_effect_size_specify(d, e)
}
\description{
Perform an effect size analysis of the timings for the standard and virtual mutation analysis techniques.
}

