% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_wilcox_rank_sum_test}
\alias{analyse_wilcox_rank_sum_test}
\title{FUNCTION: analyse_wilcox_rank_sum_test}
\usage{
analyse_wilcox_rank_sum_test(d)
}
\description{
Perform a statistical analysis of the timings for the standard and virtual mutation analysis techniques.
}

