% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/correlation.R
\name{calculate_mutation_score_correlation}
\alias{calculate_mutation_score_correlation}
\title{FUNCTION: calculate_mutation_score_correlation}
\usage{
calculate_mutation_score_correlation(data)
}
\description{
Determine the correlation between the mutation scores produced by the virtual mutation and time-constrained original
methods. Calculates Kendall's Tau_b and, additionally, returns the p-value for the hypothesis test.
}

