% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/create.R
\name{create_original_and_virtual_data}
\alias{create_original_and_virtual_data}
\title{FUNCTION: create_original_and_virtual_data}
\usage{
create_original_and_virtual_data()
}
\description{
Create a combined data file that contains both the original and the virtual data sets.
}

