% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/create.R
\name{create_original_data}
\alias{create_original_data}
\title{FUNCTION: create_original_data}
\usage{
create_original_data()
}
\description{
Create the "original" data from the HyperSQL, Postgres, and SQLite databases.
}

