% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_virtual_data}
\alias{create_virtual_data}
\title{FUNCTION: create_virtual_data}
\usage{
create_virtual_data()
}
\description{
Read virtual data from the virtual data file, then transform it and munge it into the right format.
}

