% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/read.R
\name{read_original_mutation_data}
\alias{read_original_mutation_data}
\title{FUNCTION: read_original_mutation_data}
\usage{
read_original_mutation_data()
}
\description{
Read the data file that contains the "original mutation" data. This is the data file that contains the data about
original mutation (i.e., mutation that interacts with a DBMS). Note that it contains data for more schemas than can
actually be used in the AST 2016 paper submission.
}

