% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/read.R
\name{read_original_mutation_data_hypersql_subset}
\alias{read_original_mutation_data_hypersql_subset}
\title{FUNCTION: read_original_mutation_data_hypersql_subset}
\usage{
read_original_mutation_data_hypersql_subset()
}
\description{
Read the data file that contains the "original mutation" data. This is the data file that contains the data about
original mutation (i.e., mutation that interacts with a DBMS), but only for the HyperSQL DBMS. This is the data file
that contains the correct timings for each of the individuals mutants that were run during the mutation analysis.
}

