% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_virtual_mutation_data_subset}
\alias{read_virtual_mutation_data_subset}
\title{FUNCTION: read_virtual_mutation_data_subset}
\usage{
read_virtual_mutation_data_subset()
}
\description{
Read the data file that contains the "virtual mutation" data. This is the data file that contains the data about
virtual mutation (i.e., mutation using a model of the DBMS). Data has been subset for the 16 chosen schemas for AST 2016.
Then, the data is further subset to the 9 schemas that contain a consistent mutation pipeline through all experiments.
}

