% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_original_technique}
\alias{replace_original_technique}
\title{FUNCTION: replace_original_technique}
\usage{
replace_original_technique(d)
}
\description{
Replace the original technique name in the original data set with a name that is properly descriptive.
}

