% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_virtual_technique}
\alias{replace_virtual_technique}
\title{FUNCTION: replace_virtual_technique}
\usage{
replace_virtual_technique(d)
}
\description{
Replace the virtual technique name in the original data set with a name that is capitalized.
}

