% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_all_attributes_except_identifier}
\alias{subset_all_attributes_except_identifier}
\title{FUNCTION: subset_all_attributes_except_identifier}
\usage{
subset_all_attributes_except_identifier(d)
}
\description{
Subsets out all of the attributes except for the (no longer needed) identifier attribute.
}

