% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_chosen_schemas}
\alias{subset_chosen_schemas}
\title{FUNCTION: subset_chosen_schemas}
\usage{
subset_chosen_schemas(d, m)
}
\description{
Extract only those schemas that are part of the AST 2016 paper's empirical results.
}

