% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_correct_pipeline_schemas}
\alias{subset_correct_pipeline_schemas}
\title{FUNCTION: subset_correct_pipeline_schemas}
\usage{
subset_correct_pipeline_schemas(d, v)
}
\description{
Extract only those schemas that were run with the correct mutation pipeline, for consistent results.
}

