% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_data_generator}
\alias{subset_data_generator}
\title{FUNCTION: subset_data_generator}
\usage{
subset_data_generator(d)
}
\description{
Remove any data generator results not related to the avsDefaults generator
}

