% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_virtual_attributes}
\alias{subset_virtual_attributes}
\title{FUNCTION: subset_virtual_attributes}
\usage{
subset_virtual_attributes(d)
}
\description{
Subsets out the attributes that are most relevant to the virtual mutation data analysis.
}

