% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_mutation_analysis}
\alias{summarise_mutation_analysis}
\title{FUNCTION: summarise_mutation_analysis}
\usage{
summarise_mutation_analysis(d)
}
\description{
Summarise an entire data frame (for a single DBMS) to accumulate the required data for the "original" technique.
}

