% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_execution_times_for_threshold}
\alias{transform_execution_times_for_threshold}
\title{FUNCTION: transform_execution_times_for_threshold
Transform the timing values so that they do not include any below 100 ms. This is the widely agreed on
value below which a human cannot perceive differences (see the SSBSE 2015 paper for support for this statement).}
\usage{
transform_execution_times_for_threshold(d)
}
\description{
FUNCTION: transform_execution_times_for_threshold
Transform the timing values so that they do not include any below 100 ms. This is the widely agreed on
value below which a human cannot perceive differences (see the SSBSE 2015 paper for support for this statement).
}

