% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_mutation_time_savings}
\alias{transform_mutation_time_savings}
\title{FUNCTION: transform_mutation_time_savings}
\usage{
transform_mutation_time_savings(d, o)
}
\description{
Transform the combined data file so that it has a new attribute about the savings of using virtual mutation.
}

