% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_time_constrained_data_for_scores}
\alias{transform_time_constrained_data_for_scores}
\title{FUNCTION: transform_time_constrained_data_for_scores}
\usage{
transform_time_constrained_data_for_scores(d)
}
\description{
Transform the data that contains the virtual mutation scores and the time-constrained mutation scores so that they
are "tidy" in format and thus easily amenable to visualization and analysis.
}

