% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_time_constrained_data_for_totals}
\alias{transform_time_constrained_data_for_totals}
\title{FUNCTION: transform_time_constrained_data_for_totals}
\usage{
transform_time_constrained_data_for_totals(d)
}
\description{
Transform the data that contains the virtual mutation total count and the time-constrained mutation total count so
that they are "tidy" in format and thus easily amenable to visualization and analysis.
}

