% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_truncate_execution_times_for_precision}
\alias{transform_truncate_execution_times_for_precision}
\title{FUNCTION: transform_truncate_execution_times_for_precision
Transform the timing values so that they do not include any below a specified number of ms.}
\usage{
transform_truncate_execution_times_for_precision(d, p)
}
\description{
FUNCTION: transform_truncate_execution_times_for_precision
Transform the timing values so that they do not include any below a specified number of ms.
}

