% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_virtual_mutation_time}
\alias{transform_virtual_mutation_time}
\title{FUNCTION: transform_virtual_mutation_time}
\usage{
transform_virtual_mutation_time(d)
}
\description{
Transform the virtual mutation data file so that it is has a single attribute for mutation analysis time.
}

