% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_mutation_score_time_constrained_color}
\alias{visualise_mutation_score_time_constrained_color}
\title{FUNCTION: visualise_mutation_score_time_constrained_color}
\usage{
visualise_mutation_score_time_constrained_color(data)
}
\description{
Produces a visualisation shows the number of mutants in total and killed for the virtual mutation method and for the
time-constrained original approach. Color version of the graph.
}

