% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_mutation_time_original_virtual}
\alias{visualise_mutation_time_original_virtual}
\title{FUNCTION: visualise_mutation_time_original_virtual}
\usage{
visualise_mutation_time_original_virtual(data)
}
\description{
Visualize the execution time of the original mutation analysis technique.
}

