% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_mutation_time_original_virtual_color}
\alias{visualise_mutation_time_original_virtual_color}
\title{FUNCTION: visualise_mutation_time_original_virtual_color}
\usage{
visualise_mutation_time_original_virtual_color(data)
}
\description{
Visualize the execution time of the original mutation analysis technique, now in color.
}

