% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_mutation_totals_time_constrained}
\alias{visualise_mutation_totals_time_constrained}
\title{FUNCTION: visualise_mutation_totals_time_constrained}
\usage{
visualise_mutation_totals_time_constrained(data)
}
\description{
Produces a visualisation shows the number of mutants in total and killed for the virtual mutation method and for the
time-constrained original approach
}

