% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_savings_and_mutants}
\alias{visualise_savings_and_mutants}
\title{FUNCTION: visualise_savings_and_mutants}
\usage{
visualise_savings_and_mutants(data)
}
\description{
Visualize the time savings with respect to the total number of mutants analysed.
}

