% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_savings_and_mutants_color}
\alias{visualise_savings_and_mutants_color}
\title{FUNCTION: visualise_savings_and_mutants_color}
\usage{
visualise_savings_and_mutants_color(data)
}
\description{
Visualize the time savings with respect to the total number of mutants analysed, color version.
}

