% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_savings_and_tests}
\alias{visualise_savings_and_tests}
\title{FUNCTION: visualise_savings_and_tests}
\usage{
visualise_savings_and_tests(data)
}
\description{
Visualize the time savings with respect to the total number of tests that must be run.
}

