% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_mutation_for_attributes}
\alias{rename_mutation_for_attributes}
\title{FUNCTION: rename_mutation_for_attributes}
\usage{
rename_mutation_for_attributes(d)
}
\description{
Renames the mutation data so that it uses the correct attribute names.
}

